feat: add ability to use cedarling authz before and after authentication
