+++ demos/jans-tarp/src/options/cedarlingBootstrap.json
@@ -1,0 +1,25 @@
+{
+    "CEDARLING_APPLICATION_NAME": "My App",
+    "CEDARLING_POLICY_STORE_URI": "",
+    "CEDARLING_LOG_TYPE": "std_out",
+    "CEDARLING_LOG_LEVEL": "INFO",
+    "CEDARLING_USER_AUTHZ": "disabled",
+    "CEDARLING_WORKLOAD_AUTHZ": "disabled",
+    "CEDARLING_PRINCIPAL_BOOLEAN_OPERATION": {
+        "and": [
+          { "===": [{ "var": "Jans::Workload" }, "ALLOW"] },
+          { "===": [{ "var": "Jans::User" }, "ALLOW"] }
+        ]
+      },
+    "CEDARLING_JWT_SIG_VALIDATION": "disabled",
+    "CEDARLING_JWT_STATUS_VALIDATION": "disabled",
+    "CEDARLING_MAPPING_USER": "Jans::User",
+    "CEDARLING_MAPPING_WORKLOAD": "Jans::Workload",
+    "CEDARLING_JWT_SIGNATURE_ALGORITHMS_SUPPORTED": [
+      "HS256", "RS256"
+    ],
+    "CEDARLING_ID_TOKEN_TRUST_MODE": "none",
+    "CEDARLING_LOCK": "disabled",
+    "CEDARLING_LOCK_DYNAMIC_CONFIGURATION": "disabled",
+    "CEDARLING_LOCK_LISTEN_SSE": "disabled"
+}
