+++ demos/jans-tarp/tsconfig.json
@@ -5,6 +5,7 @@
          "target": "es6",
          "moduleResolution": "node",
          "esModuleInterop": true,
+         "resolveJsonModule": true,
      },
      "include": ["src/**/*.ts", "src/**/*.tsx"],
      "exclude": ["node_modules"],
